requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.0.0
PyGithub==2.1.1
requests==2.31.0